Move reconcile logic out of the controllers package
The new operator SDK is changing the directory structure drastically.
Since we are going through the effort of moving things around it makes
sense to move most of the non-SDK related code in locations that are not
affected. That way it the migration will be easier and we shield
ourselves from having to do this again in the future if the need arises
again.