Pre SDK upgrade changes
* Remove metrics exporters

The way metrics are exported changes in the next SDK version.
Furthermore k8sutil package that's used in this code is also removed.

We are going to remove this for now and revisit after the upgrade.

* Remove deprecated packages

Some of the packages used have been replaced or removed in future SDK
versions so we are removing them.